+++ capstone/programs/capstone/src/state/config.rs
@@ -5,7 +5,7 @@ pub struct Config {
     pub admin: Pubkey,
     pub treasury: Pubkey,
     pub verifier_registry: Pubkey,
-    pub fee: u16,
+    pub fee: u64,
     pub project_count: u32,
     pub bump: u8,
 }
